fix: the 50 findings from a third adversarial review (#15)
A whole-repo adversarial review (21 areas over code, docs, charts and build)
raised 68 claims; 18 died at a refute-first gate and 50 were confirmed by an
independent reproducer. This closes all of them, plus the 8 defects three
audits of these fixes found in the fixes themselves.

The three that mattered, all on the unauthenticated ingest surface or the
kubelet path:

- OTLP nesting was unbounded. pdata's AnyValue/ArrayValue decoder recurses
  with no depth counter, so ~1.6 MiB of legal body inside every cap peaked at
  128 MiB of goroutine stack per decode and several fit inside the byte budget
  at once; nothing survived to move a counter. depth.go walks the wire shape
  before the decode (O(len), 0 allocs) on the HTTP door, the gRPC codec — the
  only hook ahead of the decoder — and the tier's internal hop. The walk is a
  strict superset of pdata's own reader: five bypasses were found and closed
  during review (a group-tag prefix, non-canonical 10-byte varints in all
  three positions, and a length-truncation mismatch), then 7.1M crafted and
  random mutations against the real decoder found no sixth.

- The ingest byte budget bounded WIRE bytes, not what they inflate into. The
  four full-size bodies it deliberately admits measured 1042 MiB of live heap
  against the 512Mi the chart ships — ~160 KiB of gzip for an OOM-kill,
  repeatable into CrashLoopBackOff. A second budget charges decoded structure,
  refusing retryably (429+Retry-After / ResourceExhausted+RetryInfo) on both
  arms. On gRPC the charge lives in the three Export methods, not the
  interceptor: grpc-go hands an interceptor pdata's internal request type, so
  an interceptor-side type switch can never match — it charged nothing for a
  full round while three tests passed by hand-building the public wrapper.

- A CRI-O conmon scope minted a second, byte-identical resource for the
  container it supervises: two conflicting samples per scrape into one
  cumulative series, so rate() fabricated a reset every cycle on every
  OpenShift node. Identity is now adopted only on positive evidence, and a
  cgroup-derived row keeps the `id` that distinguishes it.

Also: a forged k8s.namespace.name no longer picks another tenant's endpoint
(stripped at receipt, and the resolution overwrites it) — with the residual
that a stolen container.id still resolves to its owner's namespace documented
and pinned rather than papered over; the ingest log path applies logAttributes
and stages its rule-drop tally to the ack; the tailer's 1 GiB drain cap and
its node-metadata latch stop losing data silently; /v1/explain reports the
targets the per-pod ceiling refuses and no longer moves a counter it documents
itself as never moving; splitBatcher's memos are byte-bounded; native
histograms cannot ship count disagreeing with their buckets; the chart stops
excluding a namespace it sends no logs to, brackets IPv6 kubelet endpoints,
and scopes the events Role to the two objects it names.

The rest is documentation the code had outgrown — including CLAUDE.md
prescribing a gzip-body shape a fix had made memory-unsafe, and six sites that
never learned the `summary` pipeline exists.

make check, -race repo-wide, the tag-less variant and every generated artifact
are green; 24 new test files pin the fixes, each verified to fail without it.

+++ internal/agent/azurediag/auth.go
@@ -269,13 +269,23 @@ func parseTokenResponse(body []byte, what string) (string, time.Duration, error)
 			// secs > 0 (falls back to the default), but +Inf PASSES it and
 			// time.Duration(+Inf) overflows to a large NEGATIVE duration, so
 			// expiry lands in the past and the token is refreshed on every
-			// connection. Reject non-finite (and cap absurd finite) values.
-			// Non-finite AND absurd-finite: 1e300 overflows
-			// time.Duration exactly as +Inf does, landing the expiry in the
-			// past and refreshing the token on every connection — the very
-			// symptom the Inf guard was added for. maxTokenTTL is the widest
-			// value that cannot overflow the multiply.
-			const maxTokenTTL = float64(math.MaxInt64) / float64(time.Second)
+			// connection. Reject non-finite AND absurd-finite values: 1e300
+			// overflows time.Duration exactly as +Inf does, landing the expiry
+			// in the past — the very symptom the Inf guard was added for.
+			//
+			// The bound is computed in INTEGER space on purpose. Written as
+			// float64(math.MaxInt64)/float64(time.Second) the numerator rounds
+			// UP to exactly 2^63 — float64 cannot hold MaxInt64 — so the
+			// quotient was itself out of range: 9223372036.854776 passed the
+			// test and its product is exactly 2^63, which time.Duration
+			// converts to MinInt64 on amd64 (the Go spec leaves an
+			// out-of-range float→int conversion implementation-defined, so
+			// elsewhere it saturates to MaxInt64 instead — both wrong, in
+			// opposite directions). MaxInt64/int64(time.Second) = 9223372036
+			// is exact as a float64 and its product stays a comfortable
+			// margin below 2^63, so no rounding can carry it over. The
+			// rejected sliver is a TTL of ~292 years.
+			const maxTokenTTL = float64(math.MaxInt64 / int64(time.Second))
 			if secs, err := strconv.ParseFloat(s, 64); err == nil && secs > 0 && secs <= maxTokenTTL {
 				ttl = time.Duration(secs * float64(time.Second))
 			}
